requirements.txt
django-registration==2.2
httplib2==0.10.3
jsonpickle==0.9.4
lxml==3.8.0
django_extensions
Werkzeug
boardgamegeek2
